Updates to Chapter 2
Line 19: misspelled word "described"
Line 493: confused article
Line 856: this is not grammatically incorrect or anything, I just feel that adding "to find" or "to evaluate" or something along these lines makes this statement more complete as it describes what is difficult with these quantities. However, it makes the sentence stylistically uglier as it's unnecessarily long. I think it could be rephrased to: "By switching between algebraic and geometric definitions, we can use the dot product to find otherwise difficult quantities." to achieve the same "completeness" without making it longer.
Line 940: Missing comma after an introductory phrase
Line 1500: Again, not a necessary comma, but I'd consider putting it there just because it sets the stage for your statement
Line 1587: a typo