Small updates, Ch 3
Line 37: I might be wrong about this one, but I felt it needed a "has to be" because it's a requirement?]
Line 89: American or British spelling, you decide, but for consistency, yet again I'd go with the American here. Also after "time has a forward and backward", I naturally want to write a noun like "direction" or something like that, but I can't grammatically explain why because it just "feels right"
Line 146: missing comma after an introductory statement; starting to feel like it's your style of writing. It's not wrong to not write commas after these, but for all I've learned, it's stylistically better to do so
Lines 162-163: Swapped the order of the sentence because it flows more naturally this way (at least to me) (also low-key insinuated that physics can't prove what speed is, seemed unnecessary to me)
Line 397: a typo
Line 588: "in general is an interrupter, so it should be separated with commas from both sides. However, I'd just change the sentence to 
"Well, integrals are generally hard" because it's "same-same" just simpler
Line 595: American vs British spelling again
Line 729: British spelling