Changes to Ch4
Line 44: missing article
Line 221: a complex adjective
Line 222: unnecessary comma
Line 265: missing comma after an introductory statement
Lines 302 & 312: confused prepositions
Line 343: family is a singular subj
Line 385: "analogues" is a very British spelling, so I feel the urge to point it out but then again, it doesn't actually matter that much 
Lines 521&524: complex sentences are clearer with commas there. I think this paragraph is a little oddly phrased, but I can't come up with an immediate improvement, so I'll think about it later
Line 574 &577: missing articles 
Line 681: missing article
Line 843: missing comma after an introductory statement
Line 861: missing comma in a complex sentence
Line 879: let's separate the interrupter with commas here because it's a strong statement. 
Personal opinion here, the parentheses and the doubling of "used" is unnecessary, it'd look more elegant if you wrote 
"Equation \eqref{EQONEVARLINAPPROX} encapsulates such a strong idea that
it can, in fact, be used to define the derivative of a single-variable function." Here "could" is changed to "can" to imply that it can and is used for that definition.